Fix: dividing imagen Changer in little components
